refactor: move simple audio handler out of proxy
